+++ src/asldro/cli.py
@@ -1,10 +1,12 @@
 """ Command line interface (CLI) for the ASL DRO """
+import shutil
 import logging
 import argparse
 import os
 import json
 import sys
 from typing import List
+from asldro.data.filepaths import GROUND_TRUTH_DATA
 
 from asldro.examples import run_full_pipeline
 from asldro.validators.user_parameter_input import (
@@ -15,6 +17,51 @@ logging.basicConfig(
     stream=sys.stdout, format="%(asctime)s %(message)s", level=logging.INFO
 )  # Set the log level to INFO
 logger = logging.getLogger(__name__)
+
+
+class HrgtType:  # pylint: disable=too-few-public-methods
+    """A HRGT string checker.
+    Will determine if the input is a valid HRGT string label.
+    """
+
+    def __call__(self, hrgt_label: str):
+        """
+        Do the checking
+        :param hrgt_label: the HRGT label string
+        """
+        if hrgt_label not in GROUND_TRUTH_DATA.keys():
+            raise argparse.ArgumentTypeError(
+                f"{hrgt_label} is not valid, must be one of {', '.join(GROUND_TRUTH_DATA.keys())}"
+            )
+        return hrgt_label
+
+
+class DirType:  # pylint: disable=too-few-public-methods
+    """
+    A directory checker. Will determine if the input is a directory and
+    optionally, whether it exists
+    """
+
+    def __init__(self, should_exist: bool = False):
+        """
+        :param should_exist: does the directory have to exist
+        """
+        self.should_exist: bool = should_exist
+
+    def __call__(self, path: str):
+        """
+        Do the checking
+        :param path: the path to the directory
+        """
+        # Always check the file is a directory
+
+        if self.should_exist:
+            # Check whether the file exists
+            if not os.path.exists(path):
+                raise argparse.ArgumentTypeError(f"{path} does not exist")
+            if not os.path.isdir(path):
+                raise argparse.ArgumentTypeError(f"{path} is not a directory")
+        return path
 
 
 class FileType:  # pylint: disable=too-few-public-methods
@@ -93,6 +140,17 @@ def output_params(args):
         json.dump(get_example_input_params(), json_file, indent=4)
 
 
+def output_hrgt(args):
+    """Parses the 'output hrgt' subcommand. Must have a
+    'hrgt' which must be the string identifier of the hrgt and
+    'output_dir' parameter which is the directory to output
+    to."""
+    for file_name in GROUND_TRUTH_DATA[args.hrgt].values():
+        shutil.copyfile(
+            file_name, os.path.join(args.output_dir, os.path.basename(file_name))
+        )
+
+
 def main():
     """Main function for the Command Line Interface. Provides multiple options
     which are best documented by running the command line tool with `--help`"""
@@ -138,6 +196,7 @@ def main():
     )
     output_parser.set_defaults(func=lambda _: output_parser.print_help())
 
+    # Parameter output
     output_params_parser = output_type_parser.add_parser(
         name="params",
         description="""The DRO generation default params
@@ -151,6 +210,26 @@ def main():
         "Must be a JSON file. Will overwrite an existing file.",
     )
     output_params_parser.set_defaults(func=output_params)
+
+    # HRGT output
+    output_hrgt_parser = output_type_parser.add_parser(
+        name="hrgt", description="""The DRO high resolution ground truth files"""
+    )
+
+    output_hrgt_parser.add_argument(
+        "hrgt",
+        type=HrgtType(),
+        help=f"The HRGT label. One of {', '.join(GROUND_TRUTH_DATA.keys())}",
+    )
+
+    output_hrgt_parser.add_argument(
+        "output_dir",
+        type=DirType(should_exist=True),
+        help="The directory to output to. "
+        "Must exist. Will overwrite any existing HRGT files",
+    )
+
+    output_hrgt_parser.set_defaults(func=output_hrgt)
 
     generate_parser.set_defaults(func=generate)