+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Append metadata filter
 - Combine time-series images filter
 - Phase-magnitude filter
+- CLI tool to output the HRGT files
 
 ## [1.0.0] - 2020-10-16
 ### Added